+++ R/agedepth.R
@@ -131,6 +131,7 @@
 #' @param verbose Provide a summary of the age ranges after producing the age-depth model graph; default \code{verbose=FALSE}.
 #' @param roundby Rounding of the values reported at the end of the function. Defaults to 2 decimals.
 #' @param save.info By default, a variable called `info' with relevant information about the run (e.g., core name, priors, settings, ages, output) is saved into the working directory. Note that this will overwrite any existing variable with the same name - as an alternative, one could run, e.g., \code{myvar <- Bacon()}, followed by supplying the variable \code{myvar} in any subsequent commands.
+#' @param ssize Number of MCMC iterations to use.
 #' @author Maarten Blaauw, J. Andres Christen
 #' @return A plot of the age-depth model, and estimated ages incl. confidence ranges for each depth.
 #' @examples
@@ -139,12 +140,13 @@
 #'   agedepth()
 #' }
 #' @export
-agedepth <- function(set=get('info'), BCAD=set$BCAD, depth.unit=set$depth.unit, age.unit="yr", unit=depth.unit, d.lab=c(), age.lab=c(), yr.lab=age.lab, kcal=FALSE, acc.lab=c(), mem.lab=c(), d.min=c(), d.max=c(), d.by=c(), depths=set$depths, depths.file=FALSE, accordion=c(), plotatthesedepths=c(), age.min=c(), yr.min=age.min, age.max=c(), yr.max=age.max, hiatus.option=1, dark=c(), prob=set$prob, rounded=c(), d.res=400, age.res=400, yr.res=age.res, date.res=100, rotate.axes=FALSE, rev.age=FALSE, rev.yr=rev.age, rev.d=FALSE, use.raster=FALSE, flip.age=FALSE, flip.d=FALSE, maxcalc=500, height=1, calheight=1, ex=1, mirror=TRUE, up=TRUE, cutoff=.1, plot.range=TRUE, range.col=grey(.5), range.lty="12", range.lwd=1, mn.col="red", mn.lty="12", mn.lwd=1, med.col=NA, med.lty="12", med.lwd=1, C14.col=rgb(0,0,1,.35), C14.border=rgb(0,0,1,.5), cal.col=rgb(0,.5,.5,.35), cal.border=rgb(0,.5,.5,.5), dates.col=c(), pb.background=.5, pbmodelled.col=function(x) rgb(0,0,1,.5*x), pbmeasured.col="blue", pb.lim=c(), supp.col=rgb(.5,0,.5,.5), remove.tail=TRUE, MCMC.resample=TRUE, hiatus.col=grey(0.5), hiatus.lty="12", rgb.scale=c(0,0,0), rgb.res=100, slump.col=grey(0.8), normalise.dists=TRUE, same.heights=FALSE, cc=set$cc, title=set$core, title.location="topleft", title.size=1.5, plot.labels=FALSE, labels=c(), label.age=1, label.size=0.8, label.col="black", label.offset=c(0,0), label.adj=c(0.5,0), label.rot=0, after=set$after, bty="l", mar.left=c(3,3,1,.5), mar.middle=c(3,0,1,.5), mar.right=c(3,0,1,.5), mar.main=c(3,3,1,1), righthand=3, mgp=c(1.7,.7,.0), xaxs="r", yaxs="i", MCMC.col=grey(.4), post.col=grey(.8), post.border=grey(.4), prior.col=3, prior.lwd=2, prior.fontcol=2, prior.ticks="n", prior.fontsize=0.9, toppanel.fontsize=0.9, mainpanel.tickfontsize=1, mainpanel.labelfontsize=1, acc.xlim=c(), acc.ylim=c(), mem.xlim=c(), mem.ylim=c(), hiatus.xlim=c(), hiatus.ylim=c(), phi.xlim=c(), phi.ylim=c(), supp.xlim=c(), supp.ylim=c(), xaxt="s", yaxt="s", plot.pb=TRUE, pb.lty=1, plot.pdf=FALSE, dates.only=FALSE, model.only=FALSE, verbose=TRUE, roundby=2, save.info=set$save.info) {
+agedepth <- function(set=get('info'), BCAD=set$BCAD, depth.unit=set$depth.unit, age.unit="yr", unit=depth.unit, d.lab=c(), age.lab=c(), yr.lab=age.lab, kcal=FALSE, acc.lab=c(), mem.lab=c(), d.min=c(), d.max=c(), d.by=c(), depths=set$depths, depths.file=FALSE, accordion=c(), plotatthesedepths=c(), age.min=c(), yr.min=age.min, age.max=c(), yr.max=age.max, hiatus.option=1, dark=c(), prob=set$prob, rounded=c(), d.res=400, age.res=400, yr.res=age.res, date.res=100, rotate.axes=FALSE, rev.age=FALSE, rev.yr=rev.age, rev.d=FALSE, use.raster=FALSE, flip.age=FALSE, flip.d=FALSE, maxcalc=500, height=1, calheight=1, ex=1, mirror=TRUE, up=TRUE, cutoff=.1, plot.range=TRUE, range.col=grey(.5), range.lty="12", range.lwd=1, mn.col="red", mn.lty="12", mn.lwd=1, med.col=NA, med.lty="12", med.lwd=1, C14.col=rgb(0,0,1,.35), C14.border=rgb(0,0,1,.5), cal.col=rgb(0,.5,.5,.35), cal.border=rgb(0,.5,.5,.5), dates.col=c(), pb.background=.5, pbmodelled.col=function(x) rgb(0,0,1,.5*x), pbmeasured.col="blue", pb.lim=c(), supp.col=rgb(.5,0,.5,.5), remove.tail=TRUE, MCMC.resample=TRUE, hiatus.col=grey(0.5), hiatus.lty="12", rgb.scale=c(0,0,0), rgb.res=100, slump.col=grey(0.8), normalise.dists=TRUE, same.heights=FALSE, cc=set$cc, title=set$core, title.location="topleft", title.size=1.5, plot.labels=FALSE, labels=c(), label.age=1, label.size=0.8, label.col="black", label.offset=c(0,0), label.adj=c(0.5,0), label.rot=0, after=set$after, bty="l", mar.left=c(3,3,1,.5), mar.middle=c(3,0,1,.5), mar.right=c(3,0,1,.5), mar.main=c(3,3,1,1), righthand=3, mgp=c(1.7,.7,.0), xaxs="r", yaxs="i", MCMC.col=grey(.4), post.col=grey(.8), post.border=grey(.4), prior.col=3, prior.lwd=2, prior.fontcol=2, prior.ticks="n", prior.fontsize=0.9, toppanel.fontsize=0.9, mainpanel.tickfontsize=1, mainpanel.labelfontsize=1, acc.xlim=c(), acc.ylim=c(), mem.xlim=c(), mem.ylim=c(), hiatus.xlim=c(), hiatus.ylim=c(), phi.xlim=c(), phi.ylim=c(), supp.xlim=c(), supp.ylim=c(), xaxt="s", yaxt="s", plot.pb=TRUE, pb.lty=1, plot.pdf=FALSE, dates.only=FALSE, model.only=FALSE, verbose=TRUE, roundby=2, save.info=set$save.info, ssize=4e3) {
 # Load the output, if it exists
   outp <- paste0(set$prefix, ".out")
-
   if(file.exists(outp))
     set <- Bacon.AnaOut(outp, set, MCMC.resample)
+  
+  set$MCMCdiagnostics <- MCMC.diagnostics(set, ssize)
 
   # Plum-specific
   if(set$isplum) {
@@ -241,10 +243,14 @@ agedepth <- function(set=get('info'), BCAD=set$BCAD, depth.unit=set$depth.unit, 
   if(verbose)
     message("Calculating age ranges...")
   modelranges <- c()
-  ranges <- Bacon.rng(d, set, BCAD=BCAD, prob=prob)
+  if(length(rounded) == 0)
+    rounded <- ifelse(set$isplum, 1, 0) # plum tends to need higher precision
+  #ranges <- Bacon.rng(d, set, BCAD=BCAD, prob=prob)
+  ranges <- ageranges(d, paste0(set$prefix, "_ages.txt"), 
+    set=set, BCAD=BCAD, prob=prob, roundby=rounded) 
   d.rng <- d
   # calculate calendar axis limits
-  modelranges <- range(ranges[!is.na(ranges)])
+  modelranges <- range(ranges[,-1], na.rm=TRUE)
   if(length(set$calib$probs) > 0) {
     dates <- set$calib$probs
   dateranges <- c()
@@ -309,7 +315,7 @@ agedepth <- function(set=get('info'), BCAD=set$BCAD, depth.unit=set$depth.unit, 
 
   if(!dates.only) {
     if(verbose)
-      message("\nPreparing ghost graph... ")
+      message("Preparing ghost graph... ")
     agedepth.ghost(set, rotate.axes=rotate.axes, accordion=accordion, BCAD=BCAD, d.res=d.res, age.res=age.res, rev.d=rev.d, rev.age=rev.age, rgb.res=rgb.res, dark=dark, rgb.scale=rgb.scale, age.lim=age.lim, use.raster=use.raster, flip.age=flip.age, flip.d=flip.d)
   }
 
@@ -376,22 +382,25 @@ agedepth <- function(set=get('info'), BCAD=set$BCAD, depth.unit=set$depth.unit, 
     for(i in 1:ncol(th)) {
       h <- th[1,i] : th[2,i]
       if(rotate.axes) {
-        lines(ranges[h,1], d[h], col=range.col, lty=range.lty, lwd=range.lwd)
         lines(ranges[h,2], d[h], col=range.col, lty=range.lty, lwd=range.lwd)
-        lines(ranges[h,3], d[h], col=med.col, lty=med.lty, lwd=med.lwd) # median
-        lines(ranges[h,4], d[h], col=mn.col, lty=mn.lty, lwd=mn.lwd) # mean
+        lines(ranges[h,3], d[h], col=range.col, lty=range.lty, lwd=range.lwd)
+        lines(ranges[h,4], d[h], col=med.col, lty=med.lty, lwd=med.lwd) # median
+        lines(ranges[h,5], d[h], col=mn.col, lty=mn.lty, lwd=mn.lwd) # mean
       } else {
-          lines(d[h], ranges[h,1], col=range.col, lty=range.lty, lwd=range.lwd)
           lines(d[h], ranges[h,2], col=range.col, lty=range.lty, lwd=range.lwd)
-          lines(d[h], ranges[h,3], col=med.col, lty=med.lty, lwd=med.lwd) # median
-          lines(d[h], ranges[h,4], col=mn.col, lty=mn.lty, lwd=mn.lwd) # mean
+          lines(d[h], ranges[h,3], col=range.col, lty=range.lty, lwd=range.lwd)
+          lines(d[h], ranges[h,4], col=med.col, lty=med.lty, lwd=med.lwd) # median
+          lines(d[h], ranges[h,5], col=mn.col, lty=mn.lty, lwd=mn.lwd) # mean
         }
     }
 
-  if(length(rounded) == 0)
-    rounded <- ifelse(set$isplum, 1, 0)
-  set$ranges <- cbind(d, round(ranges, rounded))
-  colnames(set$ranges) <- c("depth", "min", "max", "median", "mean")
+#  if(length(rounded) == 0)
+#    rounded <- ifelse(set$isplum, 1, 0) # plum tends to need higher precision
+#  set$ranges <- cbind(d, round(ranges, rounded))
+#  colnames(set$ranges) <- c("depth", "min", "max", "median", "mean")
+#  fastwrite(ranges, paste0(set$prefix, "_ages.txt"), quote=FALSE, row.names=FALSE, sep="\t") # was write.table
+
+  set$ranges <- ranges
   if(save.info)
     assign_to_global("info", set)
 
@@ -399,11 +408,9 @@ agedepth <- function(set=get('info'), BCAD=set$BCAD, depth.unit=set$depth.unit, 
     if(names(dev.cur()) != "null device")
       export.pdf(paste0(set$prefix, ".pdf"))
 
-  fastwrite(set$ranges, paste0(set$prefix, "_ages.txt"), quote=FALSE, row.names=FALSE, sep="\t") # was write.table
-  rng <- abs(round(set$ranges[,3]-set$ranges[,2], rounded))
+  rng <- abs(ranges[,3]-ranges[,2])
   min.rng <- d[which(rng==min(rng, na.rm=TRUE))]
   max.rng <- d[which(rng==max(rng, na.rm=TRUE))]
-
   if(length(min.rng)==1)
     min.rng <- paste(age.unit, "at", min.rng, noquote(depth.unit)) else
       min.rng <- paste(age.unit, "between", min(min.rng), "and", max(min.rng), noquote(depth.unit))
@@ -413,39 +420,41 @@ agedepth <- function(set=get('info'), BCAD=set$BCAD, depth.unit=set$depth.unit, 
 
   if(verbose) {
     if(!dates.only)
-      message("\nMean ", 100*prob, "% confidence ranges ", round(mean(rng), rounded), " ", age.unit, ", min. ",
+      message("\nMean ", 100*prob, "% confidence ranges ", 
+	    round(mean(rng), rounded), " ", age.unit, ", min. ",
         min(rng), " ", min.rng, ", max. ", max(rng), " ", max.rng)
 
     if(set$isplum) {
       below <- which(bg > pb.background)
       if(length(below) == 0)
         message("\nIt seems that the Pb-measurements haven't reached background") else {
-        message("Pb-210 likely at or below detection limit from ", min(set$dets[below,4]), " ", set$depth.unit, " depth onward: ")
-        for(i in seq_along(below))
-          cat(set$dets[below[i],4], " ", set$depth.unit,
-            " (", round(100*bg[below[i]]), "%)",
-            if(i < length(below)) ", " else "\n", sep = "")
-      }
+          message("Pb-210 likely at or below detection limit from ", 
+		    min(set$dets[below,4]), " ", set$depth.unit, " depth onward: ")
+          for(i in seq_along(below))
+            cat(set$dets[below[i],4], " ", set$depth.unit,
+              " (", round(100*bg[below[i]]), "%)",
+              if(i < length(below)) ", " else "\n", sep = "")
+        }
     } else
         overlap(set)
-	
-	# report summaries of posteriors	
+
+    # report summaries of posteriors
     posteriors <- c(set$post.acc, set$post.mem)	
     if(!is.na(hiatus.depths[1]))
       posteriors <- c(posteriors, set$post.hiatus)
-	posteriors <- round(posteriors, roundby)
-	if(!is.na(hiatus.depths[1]))
-	  message("Posteriors: accrate mean ", posteriors[1], ", shape ", posteriors[2], 
-	    ", memory mean ", posteriors[3], ", strength ", posteriors[4], 
-		", hiatus mean ", posteriors[5], ", shape ", posteriors[6]) else
-  	  message("Posteriors: accrate mean ", posteriors[1], ", shape ", posteriors[2], 
-  	    ", memory mean ", posteriors[3], ", strength ", posteriors[4])
-		
+    posteriors <- round(posteriors, roundby)
+    if(!is.na(hiatus.depths[1]))
+      message("Posteriors: accrate mean ", posteriors[1], ", shape ", posteriors[2],
+        ", memory mean ", posteriors[3], ", strength ", posteriors[4],
+        ", hiatus mean ", posteriors[5], ", shape ", posteriors[6]) else
+      message("Posteriors: accrate mean ", posteriors[1], ", shape ", posteriors[2],
+        ", memory mean ", posteriors[3], ", strength ", posteriors[4])
+
     if(set$isplum) {
       plumpost <- round(c(set$post.phi, set$post.supp), roundby)
       message("phi mean ", plumpost[1], ", shape ", plumpost[2], 
-	    ", supported mean ", plumpost[3], ", shape ", plumpost[4])
-	}	
+        ", supported mean ", plumpost[3], ", shape ", plumpost[4])
+    }
   }
   
   #par(oldpar) # tmp Jan 2021